Revert "15686 - Remove Restx (#2494)"
This reverts commit 9f8a9bcbb9cd670950ed8bc7d62c47e5f9eecf78.
