+++ auth-api/src/auth_api/resources/account.py
@@ -11,37 +11,34 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Endpoints to reset test data from database."""
+"""API endpoints for managing an Org resource."""
 
+from flask import request
 from flask_restx import Namespace, Resource, cors
 
 from auth_api import status as http_status
 from auth_api.auth import jwt as _jwt
-from auth_api.exceptions import BusinessException
-from auth_api.services import ResetTestData as ResetService
+from auth_api.services.authorization import Authorization as AuthorizationService
 from auth_api.tracer import Tracer
-from auth_api.utils.roles import Role
 from auth_api.utils.util import cors_preflight
 
 
-API = Namespace('reset', description='Authentication System - Reset test data')
+API = Namespace('accounts', description='Endpoints for accounts management')
+
 TRACER = Tracer.get_instance()
 
 
-@cors_preflight('POST, PUT, OPTIONS')
-@API.route('', methods=['POST', 'PUT', 'OPTIONS'])
-class Reset(Resource):
-    """Cleanup test data by the provided token."""
+@cors_preflight('GET,OPTIONS')
+@API.route('/<int:account_id>/products/<string:product_code>/authorizations', methods=['GET', 'OPTIONS'])
+class AccountAuthorizations(Resource):
+    """Resource for returning authorizations for a product in an account."""
 
     @staticmethod
+    @_jwt.requires_auth
     @TRACER.trace()
     @cors.crossdomain(origin='*')
-    @_jwt.has_one_of_roles([Role.TESTER.value])
-    def post():
-        """Cleanup test data by the provided token."""
-        try:
-            ResetService.reset()
-            response, status = '', http_status.HTTP_204_NO_CONTENT
-        except BusinessException as exception:
-            response, status = {'code': exception.code, 'message': exception.message}, exception.status_code
-        return response, status
+    def get(account_id, product_code):
+        """Return authorizations for a product in an account."""
+        expanded: bool = request.args.get('expanded', False)
+        authorizations = AuthorizationService.get_account_authorizations_for_product(account_id, product_code, expanded)
+        return authorizations, http_status.HTTP_200_OK
