+++ auth-api/src/auth_api/resources/bcol_profiles.py
@@ -11,37 +11,40 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Endpoints to reset test data from database."""
+"""API endpoints for managing an Org resource."""
 
+from flask import request
 from flask_restx import Namespace, Resource, cors
 
-from auth_api import status as http_status
 from auth_api.auth import jwt as _jwt
 from auth_api.exceptions import BusinessException
-from auth_api.services import ResetTestData as ResetService
+from auth_api.services.org import Org
 from auth_api.tracer import Tracer
 from auth_api.utils.roles import Role
 from auth_api.utils.util import cors_preflight
 
 
-API = Namespace('reset', description='Authentication System - Reset test data')
+API = Namespace('BC Online Profiles', description='Endpoints for BC Online Profiles')
+
 TRACER = Tracer.get_instance()
 
 
-@cors_preflight('POST, PUT, OPTIONS')
-@API.route('', methods=['POST', 'PUT', 'OPTIONS'])
-class Reset(Resource):
-    """Cleanup test data by the provided token."""
+@cors_preflight('POST,OPTIONS')
+@API.route('', methods=['POST', 'OPTIONS'])
+class BcOnlineProfiles(Resource):
+    """Resource for validating BC Online account."""
 
     @staticmethod
+    @_jwt.has_one_of_roles([Role.STAFF_MANAGE_ACCOUNTS.value, Role.PUBLIC_USER.value])
     @TRACER.trace()
     @cors.crossdomain(origin='*')
-    @_jwt.has_one_of_roles([Role.TESTER.value])
     def post():
-        """Cleanup test data by the provided token."""
+        """Return BC Online profile details."""
+        request_json = request.get_json()
+
         try:
-            ResetService.reset()
-            response, status = '', http_status.HTTP_204_NO_CONTENT
+            bcol_response = Org.get_bcol_details(bcol_credential=request_json)
+            response, status = bcol_response.json(), bcol_response.status_code
         except BusinessException as exception:
             response, status = {'code': exception.code, 'message': exception.message}, exception.status_code
         return response, status
