+++ auth-api/src/auth_api/resources/bulk_user.py
@@ -11,37 +11,45 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Endpoints to reset test data from database."""
+"""API endpoints for managing a User resource."""
 
+from flask import request
 from flask_restx import Namespace, Resource, cors
 
 from auth_api import status as http_status
 from auth_api.auth import jwt as _jwt
 from auth_api.exceptions import BusinessException
-from auth_api.services import ResetTestData as ResetService
+from auth_api.schemas import utils as schema_utils
+from auth_api.services.user import User as UserService
 from auth_api.tracer import Tracer
-from auth_api.utils.roles import Role
 from auth_api.utils.util import cors_preflight
 
 
-API = Namespace('reset', description='Authentication System - Reset test data')
+API = Namespace('bulk users', description='Endpoints for bulk user profile management')
 TRACER = Tracer.get_instance()
 
 
-@cors_preflight('POST, PUT, OPTIONS')
-@API.route('', methods=['POST', 'PUT', 'OPTIONS'])
-class Reset(Resource):
-    """Cleanup test data by the provided token."""
+@cors_preflight('POST,OPTIONS')
+@API.route('', methods=['POST', 'OPTIONS'])
+class BulkUser(Resource):
+    """Resource for managing bulk  users post."""
 
     @staticmethod
     @TRACER.trace()
     @cors.crossdomain(origin='*')
-    @_jwt.has_one_of_roles([Role.TESTER.value])
+    @_jwt.requires_auth
     def post():
-        """Cleanup test data by the provided token."""
+        """Admin users can post multiple users to his org.Use it for anonymous purpose only."""
         try:
-            ResetService.reset()
-            response, status = '', http_status.HTTP_204_NO_CONTENT
+            request_json = request.get_json()
+            valid_format, errors = schema_utils.validate(request_json, 'bulk_user')
+            if not valid_format:
+                return {'message': schema_utils.serialize(errors)}, http_status.HTTP_400_BAD_REQUEST
+
+            users = UserService.create_user_and_add_membership(request_json['users'], request_json['orgId'])
+            is_any_error = any(user['http_status'] != 201 for user in users['users'])
+
+            response, status = users, http_status.HTTP_207_MULTI_STATUS if is_any_error else http_status.HTTP_200_OK
         except BusinessException as exception:
             response, status = {'code': exception.code, 'message': exception.message}, exception.status_code
         return response, status
