+++ auth-api/src/auth_api/resources/notifications.py
@@ -11,37 +11,38 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Endpoints to reset test data from database."""
+"""API endpoints for managing an Notification resource."""
 
 from flask_restx import Namespace, Resource, cors
 
 from auth_api import status as http_status
 from auth_api.auth import jwt as _jwt
 from auth_api.exceptions import BusinessException
-from auth_api.services import ResetTestData as ResetService
+from auth_api.services import Membership as MembershipService
 from auth_api.tracer import Tracer
 from auth_api.utils.roles import Role
 from auth_api.utils.util import cors_preflight
 
 
-API = Namespace('reset', description='Authentication System - Reset test data')
+API = Namespace('notifications', description='Endpoints for notification management')
 TRACER = Tracer.get_instance()
 
 
-@cors_preflight('POST, PUT, OPTIONS')
-@API.route('', methods=['POST', 'PUT', 'OPTIONS'])
-class Reset(Resource):
-    """Cleanup test data by the provided token."""
+@cors_preflight('GET,OPTIONS')
+@API.route('', methods=['GET', 'OPTIONS'])
+class Notifications(Resource):
+    """Resource for managing the notifications."""
 
     @staticmethod
     @TRACER.trace()
     @cors.crossdomain(origin='*')
-    @_jwt.has_one_of_roles([Role.TESTER.value])
-    def post():
-        """Cleanup test data by the provided token."""
+    @_jwt.has_one_of_roles([Role.SYSTEM.value, Role.STAFF.value, Role.PUBLIC_USER.value])
+    def get(user_id, org_id):  # pylint:disable=unused-argument
+        """Find the count of notification remaining.If any details invalid, it returns zero."""
         try:
-            ResetService.reset()
-            response, status = '', http_status.HTTP_204_NO_CONTENT
+            # todo use the user_id instead of jwt
+            pending_count = MembershipService.get_pending_member_count_for_org(org_id)
+            response, status = {'count': pending_count}, http_status.HTTP_200_OK
         except BusinessException as exception:
             response, status = {'code': exception.code, 'message': exception.message}, exception.status_code
         return response, status
