+++ auth-api/src/auth_api/resources/ops.py
@@ -12,31 +12,42 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """Endpoints to check and manage the health of the service."""
-from flask import Blueprint
+from flask_restx import Namespace, Resource
 from sqlalchemy import exc, text
 
 from auth_api.models import db
 
 
-bp = Blueprint('OPS', __name__, url_prefix='/ops')
+API = Namespace('OPS', description='Service - OPS checks')
 
 SQL = text('select 1')
 
 
-@bp.route('healthz', methods=['GET'])
-def get_ops_healthz():
-    """Return a JSON object stating the health of the Service and dependencies."""
-    try:
-        db.engine.execute(SQL)
-    except exc.SQLAlchemyError:
-        return {'message': 'api is down'}, 500
+@API.route('healthz')
+class Healthz(Resource):
+    """Determines if the service and required dependnecies are still working.
 
-    # made it here, so all checks passed
-    return {'message': 'api is healthy'}, 200
+    This could be thought of as a heartbeat for the service
+    """
+
+    @staticmethod
+    def get():
+        """Return a JSON object stating the health of the Service and dependencies."""
+        try:
+            db.engine.execute(SQL)
+        except exc.SQLAlchemyError:
+            return {'message': 'api is down'}, 500
+
+        # made it here, so all checks passed
+        return {'message': 'api is healthy'}, 200
 
 
-@bp.route('readyz', methods=['GET'])
-def get_ops_readyz():
-    """Return a JSON object that identifies if the service is setupAnd ready to work."""
-    # TODO: add a poll to the DB when called
-    return {'message': 'api is ready'}, 200
+@API.route('readyz')
+class Readyz(Resource):
+    """Determines if the service is ready to respond."""
+
+    @staticmethod
+    def get():
+        """Return a JSON object that identifies if the service is setupAnd ready to work."""
+        # TODO: add a poll to the DB when called
+        return {'message': 'api is ready'}, 200
