+++ auth-api/src/auth_api/resources/products.py
@@ -15,27 +15,31 @@
 
 import json
 
-from flask import Blueprint
-from flask_cors import cross_origin
+from flask_restx import Namespace, Resource, cors
 
 from auth_api import status as http_status
 from auth_api.exceptions import BusinessException
 from auth_api.services import Product as ProductService
 from auth_api.tracer import Tracer
-from auth_api.utils.endpoints_enums import EndpointEnum
+from auth_api.utils.util import cors_preflight
 
 
-bp = Blueprint('PRODUCTS', __name__, url_prefix=f'{EndpointEnum.API_V1.value}/products')
+API = Namespace('products', description='Endpoints for products management')
 TRACER = Tracer.get_instance()
 
 
-@bp.route('', methods=['GET', 'OPTIONS'])
-@TRACER.trace()
-@cross_origin(origin='*')
-def get_products():
-    """Get a list of all products."""
-    try:
-        response, status = json.dumps(ProductService.get_products()), http_status.HTTP_200_OK
-    except BusinessException as exception:
-        response, status = {'code': exception.code, 'message': exception.message}, exception.status_code
-    return response, status
+@cors_preflight('GET,OPTIONS')
+@API.route('', methods=['GET', 'OPTIONS'])
+class Products(Resource):
+    """Resource for managing products."""
+
+    @staticmethod
+    @TRACER.trace()
+    @cors.crossdomain(origin='*')
+    def get():
+        """Get a list of all products."""
+        try:
+            response, status = json.dumps(ProductService.get_products()), http_status.HTTP_200_OK
+        except BusinessException as exception:
+            response, status = {'code': exception.code, 'message': exception.message}, exception.status_code
+        return response, status
